# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hubuum REST API",
    "description": "OpenAPI documentation for the Hubuum REST service.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment base URL"
    }
  ],
  "paths": {
...
    "/api/v1/classes": {
      "get": {
        "tags": [
          "classes"
        ],
        "summary": "Get Api V1 Classes",
        "description": "Auto-generated documentation for GET /api/v1/classes. Supports cursor pagination through the `limit`, `sort`, and `cursor` query parameters. The next page cursor is returned in the `X-Next-Cursor` response header.",
        "operationId": "getApiV1Classes",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 100. Maximum is 250.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields. Cursor pagination uses the requested sort order and appends a stable tie-breaker automatically.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in the X-Next-Cursor response header from a previous page. Supply it unchanged to fetch the next page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Classes matching optional query filters The response body contains the current page items as a JSON array. Use the `X-Next-Cursor` header, when present, to request the next page.",
            "headers": {
              "X-Next-Cursor": {
                "schema": {
                  "type": "string"
                },
                "description": "Opaque cursor for the next page. This header is omitted when there are no more results."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HubuumClassExpanded"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "classes"
        ],
        "summary": "Post Api V1 Classes",
        "description": "Auto-generated documentation for POST /api/v1/classes.",
        "operationId": "postApiV1Classes",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewHubuumClass"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Class created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HubuumClassExpanded"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/classes/{class_id}": {
      "get": {
        "tags": [
          "classes"
        ],
        "summary": "Get Api V1 Classes By Class Id",
        "description": "Auto-generated documentation for GET /api/v1/classes/{class_id}.",
        "operationId": "getApiV1ClassesByClassId",
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "description": "Class ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Class",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HubuumClassExpanded"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Class not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "classes"
        ],
        "summary": "Delete Api V1 Classes By Class Id",
        "description": "Auto-generated documentation for DELETE /api/v1/classes/{class_id}.",
        "operationId": "deleteApiV1ClassesByClassId",
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "description": "Class ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Class deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Class not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "classes"
        ],
        "summary": "Patch Api V1 Classes By Class Id",
        "description": "Auto-generated documentation for PATCH /api/v1/classes/{class_id}.",
        "operationId": "patchApiV1ClassesByClassId",
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "description": "Class ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHubuumClass"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated class",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HubuumClassExpanded"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Class not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/classes/{class_id}/": {
      "get": {
        "tags": [
          "classes"
        ],
        "summary": "Get Api V1 Classes By Class Id Trailing",
        "description": "Auto-generated documentation for GET /api/v1/classes/{class_id}/. Supports cursor pagination through the `limit`, `sort`, and `cursor` query parameters. The next page cursor is returned in the `X-Next-Cursor` response header.",
        "operationId": "getApiV1ClassesByClassIdTrailing",
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "description": "Class ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 100. Maximum is 250.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields. Cursor pagination uses the requested sort order and appends a stable tie-breaker automatically.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in the X-Next-Cursor response header from a previous page. Supply it unchanged to fetch the next page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Objects in class The response body contains the current page items as a JSON array. Use the `X-Next-Cursor` header, when present, to request the next page.",
            "headers": {
              "X-Next-Cursor": {
                "schema": {
                  "type": "string"
                },
                "description": "Opaque cursor for the next page. This header is omitted when there are no more results."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HubuumObject"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Class not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "classes"
        ],
        "summary": "Post Api V1 Classes By Class Id Trailing",
        "description": "Auto-generated documentation for POST /api/v1/classes/{class_id}/.",
        "operationId": "postApiV1ClassesByClassIdTrailing",
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "description": "Class ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewHubuumObject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Object created in class",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HubuumObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/classes/{class_id}/relations/{relation_id}": {
      "delete": {
        "tags": [
          "classes"
        ],
        "summary": "Delete Api V1 Classes By Class Id Relations By Relation Id",
        "description": "Auto-generated documentation for DELETE /api/v1/classes/{class_id}/relations/{relation_id}.",
        "operationId": "deleteApiV1ClassesByClassIdRelationsByRelationId",
        "parameters": [
          {
            "name": "class_id",
            "in": "path",
            "description": "Class ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "relation_id",
            "in": "path",
            "description": "Class relation ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Class relation deleted"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Class or relation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/relations/classes": {
      "get": {
        "tags": [
          "relations"
        ],
        "summary": "Get Api V1 Relations Classes",
        "description": "Auto-generated documentation for GET /api/v1/relations/classes. Supports cursor pagination through the `limit`, `sort`, and `cursor` query parameters. The next page cursor is returned in the `X-Next-Cursor` response header.",
        "operationId": "getApiV1RelationsClasses",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return. Defaults to 100. Maximum is 250.",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields. Cursor pagination uses the requested sort order and appends a stable tie-breaker automatically.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned in the X-Next-Cursor response header from a previous page. Supply it unchanged to fetch the next page.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Class relations matching optional query filters The response body contains the current page items as a JSON array. Use the `X-Next-Cursor` header, when present, to request the next page.",
            "headers": {
              "X-Next-Cursor": {
                "schema": {
                  "type": "string"
                },
                "description": "Opaque cursor for the next page. This header is omitted when there are no more results."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HubuumClassRelation"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "relations"
        ],
        "summary": "Post Api V1 Relations Classes",
        "description": "Auto-generated documentation for POST /api/v1/relations/classes.",
        "operationId": "postApiV1RelationsClasses",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewHubuumClassRelation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Class relation created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HubuumClassRelation"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "example": {
          "error": "Unauthorized",
          "message": "Authentication failure"
        }
      },
...
      "HubuumClassExpanded": {
        "type": "object",
        "required": [
          "id",
          "name",
          "namespace",
          "validate_schema",
          "description",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "json_schema": {},
          "name": {
            "type": "string"
          },
          "namespace": {
            "$ref": "#/components/schemas/Namespace"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "validate_schema": {
            "type": "boolean"
          }
        }
      },
...
      "NewHubuumClass": {
        "type": "object",
        "required": [
          "name",
          "namespace_id",
          "description"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "json_schema": {},
          "name": {
            "type": "string"
          },
          "namespace_id": {
            "type": "integer",
            "format": "int32"
          },
          "validate_schema": {
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "example": {
          "description": "Server inventory class",
          "json_schema": null,
          "name": "server",
          "namespace_id": 1,
          "validate_schema": false
        }
      },
...
      "UpdateHubuumClass": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "json_schema": {},
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "namespace_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "validate_schema": {
            "type": [
              "boolean",
              "null"
            ]
          }
        },
        "example": {
          "description": "Validated server inventory class",
          "json_schema": null,
          "name": "server",
          "namespace_id": 1,
          "validate_schema": true
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}